requirements.txt
# Core dependencies
playwright>=1.40.0
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.0.0

# AI and LLM dependencies
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
openai>=1.0.0

# Local AI dependencies
ollama>=0.1.0
llama-cpp-python>=0.2.0
transformers>=4.35.0
torch>=2.0.0
accelerate>=0.20.0

# HTTP client for API browser
aiohttp>=3.8.0

# Utilities
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.15.0
webdriver-manager>=4.0.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Optional: LangSmith for tracing
langsmith>=0.0.50
